use interface to ensure we provide same methods for all cache implementations
